Add comments and clean up code
