+++ resources/views/sponsors/home.blade.php
@@ -5,7 +5,5 @@ Resume Book | {{ config('app.name') }}
 @endsection
 
 @section('content')
-
-<!-- <resume-book-index :users='@json($users)'></resume-book-index> -->
-<resume-book-index-test></resume-book-index-test>
+<resume-book-index></resume-book-index>
 @endsection
